+++ app/src/main/java/com/arctouch/codechallenge/model/Genre.kt
@@ -1,2 +1,3 @@
 package com.arctouch.codechallenge.model
 
+data class Genre(val id: Int, val name: String)
