+++ app/src/main/java/com/arctouch/codechallenge/model/GenreResponse.kt
@@ -1,2 +1,3 @@
 package com.arctouch.codechallenge.model
 
+data class GenreResponse(val genres: List<Genre>)
