+++ app/src/main/java/com/arctouch/codechallenge/model/Movie.kt
@@ -2,17 +2,6 @@ package com.arctouch.codechallenge.model
 
 import com.google.gson.annotations.SerializedName
 
-data class GenreResponse(val genres: List<Genre>)
-
-data class Genre(val id: Int, val name: String)
-
-data class UpcomingMoviesResponse(
-        val page: Int,
-        val results: List<Movie>,
-        @SerializedName("total_pages") val totalPages: Int,
-        @SerializedName("total_results") val totalResults: Int
-)
-
 data class Movie(
     val id: Int,
     val title: String,
